+++ src/tests/unit/test_snmp_mib_dupes.py
@@ -24,7 +24,7 @@ import os
 
 import pytest
 
-from watchfuls.snmp.mib_admin import MibAdmin as MA
+from lib.core.snmp.mibs.admin import MibAdmin as MA
 
 
 @pytest.fixture()